fix: レビュー指摘事項の修正
- end_timeに(int)キャストを追加（float暗黙変換を防止）
- バリデーション済みフィールドの冗長な ?? [] を削除
- subtitles.*.durationのmax値をSubtitleApiControllerと統一(120→60)

+++ app/Services/Highlights/HighlightCandidateExtractor.php
@@ -118,7 +118,7 @@ class HighlightCandidateExtractor
             $rawCandidates[] = [
                 'window_index' => $i,
                 'time' => $i * self::WINDOW_SEC,
-                'end_time' => min($duration, ($i + 1) * self::WINDOW_SEC),
+                'end_time' => (int) min($duration, ($i + 1) * self::WINDOW_SEC),
                 'score' => $score,
                 'volume_score' => $volumeScore,
                 'chat_score' => $chatScore,
